test(angular): split e2e tests
